requirements.txt
asgiref==3.8.1
Django==5.1.1
factory_boy==3.3.1
Faker==29.0.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
